HARP-8744 Fix directional lighting with shadows
Standard and fill techniques use MapMeshFlatStandardMaterial
instead of the MapMeshBasicMaterial, this allows shadows
with correct lighting (i.e. the final colors aren't influenced
by the lights direction (there is no diffuse component, just
the ambient (color) and the shadow color).

The brightness of the shadow can be controlled with shadowStrength

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>